refactor: relocate live rates module
